Add new helper to reduce serverless boilerplate
